requirements.txt
# GO AI Bench — Dependencies
#
# NOTE: UTMOSv2 must be installed separately:
#   pip install git+https://github.com/sarulab-speech/UTMOSv2.git

# Core ML
torch>=2.1.0
transformers>=4.40.0
datasets>=2.19.0
huggingface-hub>=0.22.0
accelerate>=0.29.0
sentencepiece==0.2.1
sacremoses>=0.1.1

# MT Metrics
sacrebleu>=2.4.0
unbabel-comet>=2.2.0

# ASR
jiwer>=3.0.3
librosa>=0.10.1
soundfile>=0.12.1
openai-whisper>=20231117

# TTS Metrics
# UTMOSv2: install from git (see above)
pymcd>=0.1.0

# Visualization
matplotlib>=3.8.0
pandas>=2.1.0

# CLI & utilities
rich>=13.7.0
tqdm>=4.66.0
pyyaml>=6.0.1
filelock>=3.13.0
click>=8.1.7
numpy>=1.26.0

# Testing
pytest>=8.1.0
pytest-cov>=5.0.0
